kernel search, possibly broken rewrite rules or codegen?
