+++ crates/luminal_2/Cargo.toml
@@ -20,3 +20,5 @@ egg = "0.9.5"
 symbolic_expressions = "5.0.3"
 rustc-hash = "2.1.1"
 rand = "0.9.1"
+egraph-serialize = "0.2.0"
+indexmap = "2.9.0"
